feat: :sparkles: re-enable jellyfin and update HCP secret app name
This commit re-enables the Jellyfin application within the Cloudflare Zero Trust configuration. The previous commit reverted this change.

Additionally, this commit updates the default value for the `hcp_secret_app_name` variable from `tf-cloudflare` to `fs-secrets`. This change ensures that the correct HCP application is used for storing secrets.

The remote backend for Terraform is also commented out.